requirements.txt
# Aero Shield v1.8 — Python Dependencies
# Install: pip install -r requirements.txt

# Core data science
pandas==2.2.1
numpy==1.26.4
protobuf==4.25.3

# HTTP / APIs
requests==2.31.0

# Environment management
python-dotenv==1.0.1

# Geospatial visualization
folium==0.16.0

# Notebook
jupyter==1.0.0
ipykernel==6.29.3

# Dashboard
streamlit==1.33.0
streamlit-folium==0.20.0
plotly==5.20.0

# AWS pipeline
boto3==1.34.69
apache-airflow==2.9.0
apache-airflow-providers-amazon==8.19.0

# Data validation
great-expectations==0.18.14

# dbt
dbt-core==1.7.3
dbt-postgres==1.7.3
dbt-redshift==1.7.3

# Testing
pytest==8.1.1
pytest-cov==5.0.0

# Code quality
black==24.3.0
flake8==7.0.0
